+++ backend/src/main/java/com/example/phoebe/dto/response/PublicNewsDto.java
@@ -3,7 +3,7 @@ package com.example.phoebe.dto.response;
 import java.time.ZonedDateTime;
 
 public class PublicNewsDto {
-    private Integer id;
+    private Long id;
     private String title;
     private String slug;
     private String teaser;
@@ -12,7 +12,7 @@ public class PublicNewsDto {
     public PublicNewsDto() {
     }
 
-    public PublicNewsDto(Integer id, String title, String slug, String teaser, ZonedDateTime publishedAt) {
+    public PublicNewsDto(Long id, String title, String slug, String teaser, ZonedDateTime publishedAt) {
         this.id = id;
         this.title = title;
         this.slug = slug;
@@ -20,11 +20,11 @@ public class PublicNewsDto {
         this.publishedAt = publishedAt;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
